cantor: oneshot engine: inline check for layer existance

+++ cantor/ru_manager.h
@@ -3,8 +3,8 @@
 #include <stdint.h>
 
 /* Depth-safe suspend/resume of a target layer (e.g., L_RUSSIAN) */
-void ru_suspend(uint8_t layer);
-void ru_resume(uint8_t layer);
-int  ru_suspended(uint8_t layer);
+void ru_suspend(int layer);
+void ru_resume(int layer);
+int  ru_suspended(int layer);
